tests/wpt/tests/css/css-writing-modes/parsing/unicode-bidi-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Writing Modes Level 3: parsing unicode-bidi with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#propdef-unicode-bidi">
<meta name="assert" content="unicode-bidi supports the full grammar 'normal | embed | isolate | bidi-override | isolate-override | plaintext'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("unicode-bidi", "normal");
test_valid_value("unicode-bidi", "embed");
test_valid_value("unicode-bidi", "isolate");
test_valid_value("unicode-bidi", "bidi-override");
test_valid_value("unicode-bidi", "isolate-override");
test_valid_value("unicode-bidi", "plaintext");
</script>
</body>
</html>
